+++ src/ResourceManager/Network/Commands.Network/VirtualNetworkGateway/GetAzureVirtualNetworkGatewayLearnedRouteCommand.cs
@@ -40,6 +40,9 @@ namespace Microsoft.Azure.Commands.Network
         [ValidateNotNullOrEmpty]
         public virtual string ResourceGroupName { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background")]
+        public SwitchParameter AsJob { get; set; }
+
         public override void Execute()
         {
             base.Execute();