+++ src/ResourceManager/Storage/Commands.Management.Storage/StorageAccount/UpdateAzureStorageAccountNetworkRuleSet.cs
@@ -118,6 +118,9 @@ namespace Microsoft.Azure.Commands.Management.Storage
         private bool isIpRuleSet = false;
         private bool isNetworkRuleSet = false;
 
+        [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background")]
+        public SwitchParameter AsJob { get; set; }
+
 
         public override void ExecuteCmdlet()
         {
@@ -133,11 +136,11 @@ namespace Microsoft.Azure.Commands.Management.Storage
                 var storageAccount = this.StorageClient.StorageAccounts.GetProperties(
                     this.ResourceGroupName,
                     this.Name);
-                StorageNetworkAcls storageACL = storageAccount.NetworkAcls;
+                NetworkRuleSet storageACL = storageAccount.NetworkRuleSet;
 
                 if (storageACL == null)
                 {
-                    storageACL = new StorageNetworkAcls();
+                    storageACL = new NetworkRuleSet();
                 }
 
                 PSNetworkRuleSet psNetworkRule = PSNetworkRuleSet.ParsePSNetworkRule(storageACL);
@@ -163,7 +166,7 @@ namespace Microsoft.Azure.Commands.Management.Storage
                 }
                 
                 StorageAccountUpdateParameters updateParameters = new StorageAccountUpdateParameters();
-                updateParameters.NetworkAcls = PSNetworkRuleSet.ParseStorageNetworkRule(psNetworkRule);
+                updateParameters.NetworkRuleSet = PSNetworkRuleSet.ParseStorageNetworkRule(psNetworkRule);
 
                 var updatedAccountResponse = this.StorageClient.StorageAccounts.Update(
                     this.ResourceGroupName,
@@ -172,7 +175,7 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
                 storageAccount = this.StorageClient.StorageAccounts.GetProperties(this.ResourceGroupName, this.Name);
 
-                WriteObject(PSNetworkRuleSet.ParsePSNetworkRule(storageAccount.NetworkAcls));
+                WriteObject(PSNetworkRuleSet.ParsePSNetworkRule(storageAccount.NetworkRuleSet));
             }
         }
     }